Fix function chain arrow operator processing (#290)
* Make => operator eqname only in ANTLR grammar

Per discussion with maintainers, adjust arrowfunctionspecifier to not
parse the arrow function chain operator as a eqname or varref or a
parenthesizedexpr, but only an eqname to fix parsing complexity. The
relevant code to support this in BuildCSTVisitor has still not been
implemented per the TODO comment.

* Fix BuildCSTVisitor context processing for #289

The index and offset for BuildCSTVisitor for arrow chain operator
context processing was off and therefore exceptions were occuring
when handling => function calls during evaluation and therefore
function calls were not instantiated the correct way. We also
added tests to validate it works now, uncovering another issue
in the use of fn:sum casting integers to decimals, to be fixed in
a subsequent commmit.

* Fix fn:sum to detect integers for return type.