Allow file or URI arguments to metaschema-cli (#298)
* Add UriUtils class for URI/Path translation

Currently, we have the metaschema-cli only support local file paths for
arguments for content inputs and outputs. This loader will support the
use of MetaschemaLoader.load() with URIs and unify local file path and
URL access (primarily https://, http://) with the same command line
tooling for similar remote and local access.

Co-Authored-By: David Waltermire <david.waltermire@gsa.gov>

* CLI commands allow files or URIs for #297

This PR adjusts the commands and supporting databind classes and ports
them to use URIs. The commands will use the new UriUtils class and the
toUri() function to use a remote URI supported as a URL and load the
model, schema, or target file. If a local file, it will convert the file
path to valid URI.

Integration tests have been added to validate this functionality works
with the update CLI commands.

* Specific error handling for different URI failures

Per discussion with @david-waltermire, update the exception handling and
associated error messages to identify if a bad domain for a remote URI
is used or a valid domain is used but a bad URL.

Co-authored-by: David Waltermire <david.waltermire@gsa.gov>

* Feedback: use buffered input stream for YAML URIs

* Feedback: remove infeasible UriUtilsTest cases

* Feedback: arg count check for AbstractValidateContentCommand

* Feedback: docs for UriUtils.toUri()

---------

Co-authored-by: David Waltermire <david.waltermire@gsa.gov>